requirements.txt
openpyxl ~= 3.1.3
psycopg2-binary ~= 2.9.9